// web/src/shared/api/generated/model/validationQueueItemResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ValidationQueueItemResult {
  id: string;
  subjectId: string;
  unitId: string;
  unitName: string;
  lessonId: string;
  lessonName: string;
  type: string;
  stem: string;
  difficulty: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  version: number | string;
  submittedAt: string;
  openedInSession: boolean;
}